requirements.txt
kafka-python
pandas
hdfs
python-dotenv